Changed Fixture inheritance from Disposable to IAsyncDisposable. Changed guids list in tests by ImmutableList. Added check for null value in Repository. TodoTaskView ctor's is now private.

+++ src/Todo.DAL/TodoTaskRepository.cs
@@ -49,8 +49,13 @@ namespace Todo.DAL
 
         public async Task<bool> RemoveAsync(Guid id)
         {
-            //var todoTaskData = new TodoTaskData { Id = id };
             var todoTaskData = await _dbContext.TodoTasks.SingleOrDefaultAsync(x => x.Id == id);
+
+            if (todoTaskData == null)
+            {
+                return false;
+            }
+
             try
             {
                 _dbContext.TodoTasks.Remove(todoTaskData);
